Finish Bid Status
