+++ desktop/apps/auction2/components/Layout.js
@@ -2,7 +2,7 @@ import AuctionBlock from 'desktop/components/react/auction_block/auction_block'
 import Banner from 'desktop/apps/auction2/components/page/Banner'
 import CommercialFilter from './filter'
 import Footer from 'desktop/apps/auction2/components/page/Footer'
-import HeaderContainer from 'desktop/apps/auction2/components/page/header'
+import AuctionInfoContainer from 'desktop/apps/auction2/components/page/auction_info'
 import MyActiveBids from 'desktop/apps/auction2/components/page/MyActiveBids'
 import PropTypes from 'prop-types'
 import React from 'react'
@@ -24,7 +24,7 @@ function Layout (props) {
       <Banner />
 
       <div className='main-layout-container responsive-layout-container'>
-        <HeaderContainer />
+        <AuctionInfoContainer />
 
         { showAssociatedAuctions &&
           <AuctionBlock
