+++ desktop/apps/auction2/components/page/auction_info/AuctionInfo.styl
@@ -1,9 +1,10 @@
-.auction2-header
+.auction2-auction-info
   clearfix()
 
   +respond-mobile()
-    display flex
-    align-items center
+    &__container
+      display flex
+      align-items center
 
     &_expandToFullScreen
       align-items flex-start
@@ -101,7 +102,11 @@
 
     +respond-mobile()
       garamond-size('s-body')
-      margin 10px 0 30px 0
+      margin 10px 0 20px 0
+
+  &__date-time
+    +respond-mobile()
+      margin-bottom 25px
 
   &__register-button
     margin-bottom 30px
