+++ desktop/apps/auction2/stylesheets/index.styl
@@ -22,8 +22,8 @@
 @require '../components/filter/RangeSlider'
 @require '../components/filter/WorksByFollowedArtists'
 
-@require '../components/page/header/index'
-@require '../components/page/header/Registration'
+@require '../components/page/auction_info/AuctionInfo'
+@require '../components/page/auction_info/Registration'
 @require '../components/page/Banner'
 @require '../components/page/Footer'
 @require '../components/page/MyActiveBids'
